move fixture dependencies into plugins
