--- tests/grippeweb/conftest.py
@@ -1,29 +1,0 @@
-from typing import TypeVar
-
-import pytest
-from pydantic import BaseModel
-
-ModelT = TypeVar("ModelT", bound=BaseModel)
-
-
-@pytest.fixture
-def mocked_sql_tables() -> dict[str, list[BaseModel]]:
-    return {
-        "vActualQuestion": {
-            "Id": ["AAA", "BBB"],
-            "StartedOn": ["2023-11-01 00:00:00.0000000", "2023-12-01 00:00:00.0000000"],
-            "FinishedOn": [
-                "2023-12-01 00:00:00.0000000",
-                "2024-01-01 00:00:00.0000000",
-            ],
-            "RepeatAfterDays": ["1", "2"],
-        },
-        "vWeeklyResponsesMEx": {
-            "GuidTeilnehmer": [None, None],
-            "Haushalt_Registrierer": [None, None],
-        },
-        "vMasterDataMEx": {
-            "GuidTeilnehmer": [None, None],
-            "Haushalt_Registrierer": [None, None],
-        },
-    }
