+++ tests/grippeweb/mocked_grippeweb.py
@@ -8,9 +8,7 @@ from mex.grippeweb.connector import GrippewebConnector
 
 
 @pytest.fixture
-def mocked_grippeweb(
-    mocked_sql_tables: dict[str, list[BaseModel]], monkeypatch: MonkeyPatch
-) -> None:
+def mocked_grippeweb(mocked_grippeweb_sql_tables, monkeypatch: MonkeyPatch) -> None:
     """Mock grippeweb connector."""
 
     def mocked_init(self: GrippewebConnector) -> None:
@@ -24,5 +22,28 @@ def mocked_grippeweb(
     monkeypatch.setattr(
         GrippewebConnector,
         "parse_columns_by_column_name",
-        lambda self, model: mocked_sql_tables[model],
+        lambda self, model: mocked_grippeweb_sql_tables[model],
     )
+
+
+@pytest.fixture
+def mocked_grippeweb_sql_tables() -> dict[str, list[BaseModel]]:
+    return {
+        "vActualQuestion": {
+            "Id": ["AAA", "BBB"],
+            "StartedOn": ["2023-11-01 00:00:00.0000000", "2023-12-01 00:00:00.0000000"],
+            "FinishedOn": [
+                "2023-12-01 00:00:00.0000000",
+                "2024-01-01 00:00:00.0000000",
+            ],
+            "RepeatAfterDays": ["1", "2"],
+        },
+        "vWeeklyResponsesMEx": {
+            "GuidTeilnehmer": [None, None],
+            "Haushalt_Registrierer": [None, None],
+        },
+        "vMasterDataMEx": {
+            "GuidTeilnehmer": [None, None],
+            "Haushalt_Registrierer": [None, None],
+        },
+    }
